Move bridger tests into bridgers folder
