+++ test/xchainGauges/bridgers/arbitrumBridger.ts
@@ -2,16 +2,16 @@ import { setNextBlockBaseFeePerGas } from "@nomicfoundation/hardhat-network-help
 import chai from "chai"
 import { Signer } from "ethers"
 import { deployments, ethers, network } from "hardhat"
-import { ArbitrumBridger, SDL } from "../../build/typechain"
-import { MULTISIG_ADDRESSES } from "../../utils/accounts"
-import { ALCHEMY_BASE_URL, CHAIN_ID } from "../../utils/network"
+import { ArbitrumBridger, SDL } from "../../../build/typechain"
+import { MULTISIG_ADDRESSES } from "../../../utils/accounts"
+import { ALCHEMY_BASE_URL, CHAIN_ID } from "../../../utils/network"
 import {
   BIG_NUMBER_1E18,
   getWithName,
   impersonateAccount,
   MAX_UINT256,
   setEtherBalance,
-} from "../testUtils"
+} from "../../testUtils"
 
 const { expect } = chai
